Move `common_utils.py` into `responsibleai_dashboard` directory
