+++ raiwidgets/tests/responsibleai_dashboard/test_responsibleai_dashboard_input_erroranalysis.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation
 # Licensed under the MIT License.
 import pytest
-from common_utils import CheckResponsibleAIDashboardInputTestResult
+from .common_utils import CheckResponsibleAIDashboardInputTestResult
 
 from raiwidgets.interfaces import WidgetRequestResponseConstants
 from raiwidgets.responsibleai_dashboard_input import \
